+++ livedata-validateby-kt/src/main/java/com/zhuinden/livedatavalidatebykt/LiveDataValidateBy.kt
@@ -17,67 +17,28 @@ package com.zhuinden.livedatavalidatebykt
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
+import androidx.lifecycle.Observer
 import androidx.lifecycle.Transformations
-import androidx.lifecycle.map
 import com.zhuinden.livedatacombinetuplekt.combineTuple
-import com.zhuinden.tupleskt.*
 
-fun validateBy(f1: LiveData<Boolean>, f2: LiveData<Boolean>): LiveData<Boolean> = Transformations.map(combineTuple(f1, f2)) { (t1, t2) ->
-    (t1 ?: false) && (t2 ?: false)
-}
+fun validateBy(vararg liveDatas: LiveData<Boolean>): LiveData<Boolean> = MediatorLiveData<Boolean>().also { mediator ->
+    mediator.value = liveDatas.all { it.value == true }
 
-fun validateBy(f1: LiveData<Boolean>, f2: LiveData<Boolean>, f3: LiveData<Boolean>): LiveData<Boolean> = Transformations.map(combineTuple(f1, f2, f3)) { (t1, t2, t3) ->
-    (t1 ?: false) && (t2 ?: false) && (t3 ?: false)
-}
+    for (current in liveDatas) {
+        mediator.addSource(current, Observer { valid ->
+            var isValid = valid
+            if (isValid) {
+                for (liveData in liveDatas) {
+                    if (liveData !== current) {
+                        if (liveData.value != true) {
+                            isValid = false
+                            break
+                        }
+                    }
+                }
+            }
 
-fun validateBy(f1: LiveData<Boolean>, f2: LiveData<Boolean>, f3: LiveData<Boolean>, f4: LiveData<Boolean>): LiveData<Boolean> = Transformations.map(combineTuple(f1, f2, f3, f4)) { (t1, t2, t3, t4) ->
-    (t1 ?: false) && (t2 ?: false) && (t3 ?: false) && (t4 ?: false)
-}
-
-fun validateBy(f1: LiveData<Boolean>, f2: LiveData<Boolean>, f3: LiveData<Boolean>, f4: LiveData<Boolean>, f5: LiveData<Boolean>): LiveData<Boolean> = Transformations.map(combineTuple(f1, f2, f3, f4, f5)) { (t1, t2, t3, t4, t5) ->
-    (t1 ?: false) && (t2 ?: false) && (t3 ?: false) && (t4 ?: false) && (t5 ?: false)
-}
-
-fun validateBy(f1: LiveData<Boolean>, f2: LiveData<Boolean>, f3: LiveData<Boolean>, f4: LiveData<Boolean>, f5: LiveData<Boolean>, f6: LiveData<Boolean>): LiveData<Boolean> = Transformations.map(combineTuple(f1, f2, f3, f4, f5, f6)) { (t1, t2, t3, t4, t5, t6) ->
-    (t1 ?: false) && (t2 ?: false) && (t3 ?: false) && (t4 ?: false) && (t5 ?: false) && (t6 ?: false)
-}
-
-fun validateBy(f1: LiveData<Boolean>, f2: LiveData<Boolean>, f3: LiveData<Boolean>, f4: LiveData<Boolean>, f5: LiveData<Boolean>, f6: LiveData<Boolean>, f7: LiveData<Boolean>): LiveData<Boolean>  = Transformations.map(combineTuple(f1, f2, f3, f4, f5, f6, f7)) { (t1, t2, t3, t4, t5, t6, t7) ->
-    (t1 ?: false) && (t2 ?: false) && (t3 ?: false) && (t4 ?: false) && (t5 ?: false) && (t6 ?: false) && (t7 ?: false)
-}
-
-fun validateBy(f1: LiveData<Boolean>, f2: LiveData<Boolean>, f3: LiveData<Boolean>, f4: LiveData<Boolean>, f5: LiveData<Boolean>, f6: LiveData<Boolean>, f7: LiveData<Boolean>, f8: LiveData<Boolean>): LiveData<Boolean>  = Transformations.map(combineTuple(f1, f2, f3, f4, f5, f6, f7, f8)) { (t1, t2, t3, t4, t5, t6, t7, t8) ->
-    (t1 ?: false) && (t2 ?: false) && (t3 ?: false) && (t4 ?: false) && (t5 ?: false) && (t6 ?: false) && (t7 ?: false) && (t8 ?: false)
-}
-
-fun validateBy(f1: LiveData<Boolean>, f2: LiveData<Boolean>, f3: LiveData<Boolean>, f4: LiveData<Boolean>, f5: LiveData<Boolean>, f6: LiveData<Boolean>, f7: LiveData<Boolean>, f8: LiveData<Boolean>, f9: LiveData<Boolean>): LiveData<Boolean> = Transformations.map(combineTuple(f1, f2, f3, f4, f5, f6, f7, f8, f9)) { (t1, t2, t3, t4, t5, t6, t7, t8, t9) ->
-    (t1 ?: false) && (t2 ?: false) && (t3 ?: false) && (t4 ?: false) && (t5 ?: false) && (t6 ?: false) && (t7 ?: false) && (t8 ?: false) && (t9 ?: false)
-}
-
-fun validateBy(f1: LiveData<Boolean>, f2: LiveData<Boolean>, f3: LiveData<Boolean>, f4: LiveData<Boolean>, f5: LiveData<Boolean>, f6: LiveData<Boolean>, f7: LiveData<Boolean>, f8: LiveData<Boolean>, f9: LiveData<Boolean>, f10: LiveData<Boolean>): LiveData<Boolean> = Transformations.map(combineTuple(f1, f2, f3, f4, f5, f6, f7, f8, f9, f10)) { (t1, t2, t3, t4, t5, t6, t7, t8, t9, t10) ->
-    (t1 ?: false) && (t2 ?: false) && (t3 ?: false) && (t4 ?: false) && (t5 ?: false) && (t6 ?: false) && (t7 ?: false) && (t8 ?: false) && (t9 ?: false) && (t10 ?: false)
-}
-
-fun validateBy(f1: LiveData<Boolean>, f2: LiveData<Boolean>, f3: LiveData<Boolean>, f4: LiveData<Boolean>, f5: LiveData<Boolean>, f6: LiveData<Boolean>, f7: LiveData<Boolean>, f8: LiveData<Boolean>, f9: LiveData<Boolean>, f10: LiveData<Boolean>, f11: LiveData<Boolean>): LiveData<Boolean> = Transformations.map(combineTuple(f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11)) { (t1, t2, t3, t4, t5, t6, t7, t8, t9, t10, t11) ->
-    (t1 ?: false) && (t2 ?: false) && (t3 ?: false) && (t4 ?: false) && (t5 ?: false) && (t6 ?: false) && (t7 ?: false) && (t8 ?: false) && (t9 ?: false) && (t10 ?: false) && (t11 ?: false)
-}
-
-fun validateBy(f1: LiveData<Boolean>, f2: LiveData<Boolean>, f3: LiveData<Boolean>, f4: LiveData<Boolean>, f5: LiveData<Boolean>, f6: LiveData<Boolean>, f7: LiveData<Boolean>, f8: LiveData<Boolean>, f9: LiveData<Boolean>, f10: LiveData<Boolean>, f11: LiveData<Boolean>, f12: LiveData<Boolean>): LiveData<Boolean> = Transformations.map(combineTuple(f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12)) { (t1, t2, t3, t4, t5, t6, t7, t8, t9, t10, t11, t12) ->
-    (t1 ?: false) && (t2 ?: false) && (t3 ?: false) && (t4 ?: false) && (t5 ?: false) && (t6 ?: false) && (t7 ?: false) && (t8 ?: false) && (t9 ?: false) && (t10 ?: false) && (t11 ?: false) && (t12 ?: false)
-}
-
-fun validateBy(f1: LiveData<Boolean>, f2: LiveData<Boolean>, f3: LiveData<Boolean>, f4: LiveData<Boolean>, f5: LiveData<Boolean>, f6: LiveData<Boolean>, f7: LiveData<Boolean>, f8: LiveData<Boolean>, f9: LiveData<Boolean>, f10: LiveData<Boolean>, f11: LiveData<Boolean>, f12: LiveData<Boolean>, f13: LiveData<Boolean>): LiveData<Boolean> = Transformations.map(combineTuple(f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13)) { (t1, t2, t3, t4, t5, t6, t7, t8, t9, t10, t11, t12, t13) ->
-    (t1 ?: false) && (t2 ?: false) && (t3 ?: false) && (t4 ?: false) && (t5 ?: false) && (t6 ?: false) && (t7 ?: false) && (t8 ?: false) && (t9 ?: false) && (t10 ?: false) && (t11 ?: false) && (t12 ?: false) && (t13 ?: false)
-}
-
-fun validateBy(f1: LiveData<Boolean>, f2: LiveData<Boolean>, f3: LiveData<Boolean>, f4: LiveData<Boolean>, f5: LiveData<Boolean>, f6: LiveData<Boolean>, f7: LiveData<Boolean>, f8: LiveData<Boolean>, f9: LiveData<Boolean>, f10: LiveData<Boolean>, f11: LiveData<Boolean>, f12: LiveData<Boolean>, f13: LiveData<Boolean>, f14: LiveData<Boolean>): LiveData<Boolean> = Transformations.map(combineTuple(f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14)) { (t1, t2, t3, t4, t5, t6, t7, t8, t9, t10, t11, t12, t13, t14) ->
-    (t1 ?: false) && (t2 ?: false) && (t3 ?: false) && (t4 ?: false) && (t5 ?: false) && (t6 ?: false) && (t7 ?: false) && (t8 ?: false) && (t9 ?: false) && (t10 ?: false) && (t11 ?: false) && (t12 ?: false) && (t13 ?: false) && (t14 ?: false)
-}
-
-fun validateBy(f1: LiveData<Boolean>, f2: LiveData<Boolean>, f3: LiveData<Boolean>, f4: LiveData<Boolean>, f5: LiveData<Boolean>, f6: LiveData<Boolean>, f7: LiveData<Boolean>, f8: LiveData<Boolean>, f9: LiveData<Boolean>, f10: LiveData<Boolean>, f11: LiveData<Boolean>, f12: LiveData<Boolean>, f13: LiveData<Boolean>, f14: LiveData<Boolean>, f15: LiveData<Boolean>): LiveData<Boolean> = Transformations.map(combineTuple(f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15)) { (t1, t2, t3, t4, t5, t6, t7, t8, t9, t10, t11, t12, t13, t14, t15) ->
-    (t1 ?: false) && (t2 ?: false) && (t3 ?: false) && (t4 ?: false) && (t5 ?: false) && (t6 ?: false) && (t7 ?: false) && (t8 ?: false) && (t9 ?: false) && (t10 ?: false) && (t11 ?: false) && (t12 ?: false) && (t13 ?: false) && (t14 ?: false) && (t15 ?: false)
-}
-
-fun validateBy(f1: LiveData<Boolean>, f2: LiveData<Boolean>, f3: LiveData<Boolean>, f4: LiveData<Boolean>, f5: LiveData<Boolean>, f6: LiveData<Boolean>, f7: LiveData<Boolean>, f8: LiveData<Boolean>, f9: LiveData<Boolean>, f10: LiveData<Boolean>, f11: LiveData<Boolean>, f12: LiveData<Boolean>, f13: LiveData<Boolean>, f14: LiveData<Boolean>, f15: LiveData<Boolean>, f16: LiveData<Boolean>): LiveData<Boolean> = Transformations.map(combineTuple(f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11, f12, f13, f14, f15, f16)) { (t1, t2, t3, t4, t5, t6, t7, t8, t9, t10, t11, t12, t13, t14, t15, t16) ->
-    (t1 ?: false) && (t2 ?: false) && (t3 ?: false) && (t4 ?: false) && (t5 ?: false) && (t6 ?: false) && (t7 ?: false) && (t8 ?: false) && (t9 ?: false) && (t10 ?: false) && (t11 ?: false) && (t12 ?: false) && (t13 ?: false) && (t14 ?: false) && (t15 ?: false) && (t16 ?: false)
+            mediator.value = isValid
+        })
+    }
 }