+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+-LiveData-ValidateBy-KT 2.0.0 (2020-12-09)
+--------------------------------
+
+- Replace 2-16 arity methods with vararg method.
+
 -LiveData-ValidateBy-KT 1.0.0 (2020-09-18)
 --------------------------------